% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{analyze_ngrams}
\alias{analyze_ngrams}
\title{Analyze NGrams}
\usage{
analyze_ngrams(in_text, ngram = 1, top_rows = 25)
}
\arguments{
\item{in_text}{a character vector. Text to be analyzed as a character vector.}

\item{ngram}{a numeric_vector of length 1. Ngram = 1, 2 or 3.}

\item{top_rows}{a numeric vector of length 1. Number of top rows to be
returned.}
}
\description{
Analyze text with ngram among 1, 2 or 3.
}
\details{
analyze_ngrams
}
\examples{
analyze_ngrams(in_text=c("How much wood would a woodchuck chuck if a
woodchuck could chuck wood?"))
}
\author{
Ravindra Pushker
}
