% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{analyze_trigrams}
\alias{analyze_trigrams}
\title{Analyze Trigrams}
\usage{
analyze_trigrams(in_text, top_rows = 25)
}
\arguments{
\item{in_text}{a character vector. Text to be analyzed as a character vector.}

\item{top_rows}{a numeric vector of length 1. Number of top rows to be
returned.}
}
\description{
Analyze text with ngram=3 (trigrams).
}
\details{
analyze_trigrams
}
\examples{
analyze_trigrams(in_text=c("How much wood would a woodchuck chuck if a
woodchuck could chuck wood?"))
}
\author{
Ravindra Pushker
}
