% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{analyzeNGrams}
\alias{analyzeNGrams}
\title{Analyze NGrams}
\usage{
analyzeNGrams(in_text, n = 1, top_rows = 25)
}
\arguments{
\item{in_text}{a character vector. Text to be analyzed as a character vector.}

\item{n}{a numeric_vector of length 1. Ngram = 1, 2 or 3.}

\item{top_rows}{a numeric vector of length 1. Number of top rows to be returned.}
}
\description{
Analyze text with ngram among 1, 2 or 3.
}
\details{
analyzeNGrams
}
\examples{
\dontrun{
analyzeTokens(text=c("Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.", "Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."))
}
}
\author{
Ravindra Pushker
}
