% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams.R
\name{plot_ngrams}
\alias{plot_ngrams}
\title{Plot Ngrams}
\usage{
plot_ngrams(ngrams_data, top_rows = 25, plot_nrows = 25)
}
\arguments{
\item{ngrams_data}{a data.frame containing word and n columns.}

\item{top_rows}{a numeric vector of length 1. Number of top rows to be
returned.}

\item{plot_nrows}{a numeric vector of length 1. Number of rows to be plotted.}
}
\value{
A ggplot plot object of bar chart with words and their counts.
}
\description{
Plot ngrams - Word(s) vs. Count.
}
\details{
plot_ngrams
}
\examples{
plot_ngrams(data.frame(word=c("test1", "test2"), n=c(25, 30)))
}
\author{
Ravindra Pushker
}
