% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_words.R
\name{plot_top_words}
\alias{plot_top_words}
\title{Plots the most frequently used words in a given text using word_freq()}
\usage{
plot_top_words(text, top_n = 10, exclude = NULL)
}
\arguments{
\item{text}{A string representing the input text.}

\item{top_n}{Number of top words to display (default = 10).}

\item{exclude}{Optional. A character vector of words that should be excluded from the result.}
}
\value{
A ggplot2 plot.
}
\description{
This function uses word_freq() function to calculate the word frequencies, then
displays the top N words in a bar plot.
}
\examples{
plot_top_words("This is a beautiful day. A day where you can do a lot and just relax.", top_n = 5)
}
