% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TextAnalyzer.R
\name{run_TextAnalyzer}
\alias{run_TextAnalyzer}
\title{Launch the TextAnalyzer shiny app}
\usage{
The app open the browser automatically.

run_TextAnalyzer()

}
\description{
Launches the app to analyze text data with multiples techniques.
\code{run_TextAnalyzer}
}
\examples{
 \dontrun{
 # launch the app

  run_TextAnalyzer()
  }
}
