% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_freq.R
\name{word_freq}
\alias{word_freq}
\title{Calculates frequencies of words in a text}
\usage{
word_freq(text, exclude = NULL)
}
\arguments{
\item{text}{A string representing the input text.}

\item{exclude}{Optional. A character vector of words that should be excluded form the result.}
}
\value{
A table of word frequencies, sorted in descending order.
}
\description{
This function splits the input text into words, converts them to lowercase, optionally
removes user-defined words from the counting for frequency,
and returns their frequency order in descending order.
}
\examples{
word_freq("Today is a good day! A very good day.")
word_freq("We are having fun today, and we are outside today and we are going swimming.", exclude = c("are", "and"))
}
